Add duration to track look-up, DJ cooldown
In certain cases, the track look-ups will return too many results. To
account for this, we can include the track duration in the search to
limit the number of hits and improve our chances of finding a match.

The DJ cooldown allows us to prevent writes to the playlist if a live
DJ has recently written a track to the playlist. This represents a
component of the on/off switch to prevent automation from capturing
tracks while a DJ is simulataneously on the air.