Include additional input params from StationPlaylist
We're finding that some tracks are not being sent with an artist
field. To account for this, we'll allow StationPlaylist to also
send over the "album artist" field with the hopes that at least one of
the fields will be populated.

Additionally, we've added an option to include the album year with the
request to further assist with locating the correct track.